% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_info.R
\docType{data}
\name{enhSel}
\alias{enhSel}
\title{Enhancer signals subset detected from medulloblatoma samples}
\format{a data.frame instance}
\usage{
enhSel
}
\value{
NULL, but makes available the dataframe
}
\description{
This data.frame contains 65 selected in chr15  normalized
enhancers signals subset from 25 medulloblastoma samples.
}
