% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_info.R
\docType{data}
\name{enhSelGR}
\alias{enhSelGR}
\title{Genomic coordiantes of enhancer signals subet}
\format{a GRanges object}
\usage{
enhSelGR
}
\value{
NULL, but makes available the dataset
}
\description{
This GRanges object contains the coordinates of 65
medulloblastoma enhancer signals in chr15 target region
}
