% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InTADSig.R
\docType{methods}
\name{exprs,InTADSig-method}
\alias{exprs,InTADSig-method}
\title{Gene expression counts table}
\usage{
\S4method{exprs}{InTADSig}(object)
}
\arguments{
\item{object}{InTADSig object with signals and genes}
}
\value{
Gene expression table
}
\description{
This funcion returns gene expression counts table
}
\examples{
inTadSig <- newSigInTAD(enhSel, enhSelGR, rpkmCountsSel, txsSel)
head(exprs(inTadSig))
}
