% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSignalsWithLoops.R
\name{findCorFromLoops}
\alias{findCorFromLoops}
\title{Function to perfrom correlation analysis via loops.}
\usage{
findCorFromLoops(object, method = "pearson", adj.pval = FALSE)
}
\arguments{
\item{object}{InTADSig object with signals and genes combined via loops}

\item{method}{Correlation method: "pearson" (default), "kendall", "spearman"}

\item{adj.pval}{Perform p-value adjsutment and include q-values in result}
}
\value{
A table with correlation values for signal-gene pairs including
correlation p-value and euclidian distance.
}
\description{
This function combines genes and signals using obtained loop connections.
}
