% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSignalsInTAD.R
\name{fnSE}
\alias{fnSE}
\title{Preparation for correlation analysis for a signal}
\usage{
fnSE(id, sigList, tadGR, tss, pickMaxOvlp, nearestTad)
}
\arguments{
\item{id}{Id of signal from the list}

\item{sigList}{List of signal GRs and their names}

\item{tadGR}{TAD genomic regions}

\item{tss}{Gene transcription start sites}

\item{pickMaxOvlp}{Use TAD with max overlap}

\item{nearestTad}{The table listing TADs nearest to each TSS
#'}
}
\value{
Data.frame containing genes connected to signal
}
\description{
This function collects all genes for signal genomic region inside of
Topologically Associated Domains (TADs)
}
\details{
The signal is checked if it is lying inside of TAD.
Then all genes in this TAD are collected.
}
