% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InTADSig.R
\docType{methods}
\name{geneCoords}
\alias{geneCoords}
\alias{geneCoords,InTADSig-method}
\title{Gene coords GRanges}
\usage{
geneCoords(object)

\S4method{geneCoords}{InTADSig}(object)
}
\arguments{
\item{object}{InTADSig object with signals and genes}
}
\value{
Gene GRanges
}
\description{
This funcion returns the gene GRanges
}
\examples{
inTadSig <- newSigInTAD(enhSel, enhSelGR, rpkmCountsSel, txsSel)
head(geneCoords(inTadSig))
}
