% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGeneExpr.R
\name{get.enr.bg.normfit}
\alias{get.enr.bg.normfit}
\title{Function to estimate gene expression}
\usage{
get.enr.bg.normfit(x)
}
\arguments{
\item{x}{Full gene expression vector}
}
\value{
Distribution properties: mean and std
}
\description{
This function uses mclust package to analyze gene expression distribution
}
\details{
The function adjust filtering cut value based on mclust library to exclude
low expressed genes. It is a part of filtering procedure.
}
