% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InTADSig.R
\docType{class}
\name{InTADSig}
\alias{InTADSig}
\alias{InTADSig-class}
\title{The InTADSig Class}
\description{
The InTADSig object stores signals and gene expression data for the samples.
}
\details{
It uses MultiAssayExperiment object to store information. Key slots to access
are listed below.
}
\section{Slots}{

   \describe{
       \item{\code{sigMAE}:}{\code{"MultiAssayExperiment"},
       MultiAssayExperiment object containg signals and gene counts }
       \item{\code{signalConnections}:}{\code{"list"},
       The list of signals representing gene data frames in the same TAD }
       \item{\code{ncore}:}{\code{"numeric"},
       Number of cores to use for parallel computing }#'
}
}

