% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_info.R
\docType{data}
\name{loopsDfSel}
\alias{loopsDfSel}
\title{Data frame containing coordinates of loops}
\format{a data.frame object}
\usage{
loopsDfSel
}
\value{
NULL, but makes available the dataset
}
\description{
The table contains genomic coordinates of chromatin loops
in 6-column format derived from IMR90 cell line (focus : chr15)
}
