% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_info.R
\docType{data}
\name{mbAnnData}
\alias{mbAnnData}
\title{Data frame containing information about samples}
\format{a data.frame object}
\usage{
mbAnnData
}
\value{
NULL, but makes available the dataset
}
\description{
The table includes additional informaiton about MB tumour
samples (subgroup, gender, age, histology and M.Stage)
}
