% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InTADSig-methods.R
\name{newSigInTAD}
\alias{newSigInTAD}
\title{Create InTADSig object}
\usage{
newSigInTAD(signalData = NULL, signalRegions = NULL, countsData = NULL,
  geneRegions = NULL, sampleInfo = NULL, performLog = TRUE,
  logExprsOffset = 1, ncores = 1)
}
\arguments{
\item{signalData}{data frame containing signals}

\item{signalRegions}{genomic regions of the signals}

\item{countsData}{data matrix containing count expression values}

\item{geneRegions}{gene coordiantes}

\item{sampleInfo}{data frame containing additional sample info}

\item{performLog}{Perform log2 convertion of expression values.
Default: TRUE.}

\item{logExprsOffset}{Offset x for log2 gene exrpression
i.e. log2(value + x). Default: 1}

\item{ncores}{Number of cores to use for parallel computing}
}
\value{
Novel InTADSig object
}
\description{
The fuction generates an object that contains the signals and gene expression
data.frames along with their genomic coordinates for further processing.
}
\details{
InTADSig object stores matrices of signals and gene expression values along
with coordinates. The order of samples and names of columns should match in both
datasets. For gene coordinates GRanges "gene_id" and "gene_name" are required
in metadata. These are typical markers of genes in GTF anntotation format.
}
\examples{
## create sigInTAD object
inTadSig <- newSigInTAD(enhSel, enhSelGR, rpkmCountsSel, txsSel)


}
