% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_info.R
\docType{data}
\name{rpkmCountsSel}
\alias{rpkmCountsSel}
\title{Gene expression subset from medulloblastoma samples}
\format{a data.frame instance}
\usage{
rpkmCountsSel
}
\value{
NULL, but makes available the dataframe
}
\description{
This data.frame contains RPKM gene expression values from chr15
for subset from 25 medulluoblastoma samples.
}
