% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InTADSig.R
\docType{methods}
\name{sigCoords}
\alias{sigCoords}
\alias{sigCoords,InTADSig-method}
\title{Signal coords GRanges}
\usage{
sigCoords(object)

\S4method{sigCoords}{InTADSig}(object)
}
\arguments{
\item{object}{InTADSig object with signals and genes}
}
\value{
Signal GRanges
}
\description{
This funcion returns the signal GRanges
}
\examples{
inTadSig <- newSigInTAD(enhSel, enhSelGR, rpkmCountsSel, txsSel)
head(sigCoords(inTadSig))
}
