% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InTADSig.R
\docType{methods}
\name{signals}
\alias{signals}
\alias{signals,InTADSig-method}
\title{Signal values table}
\usage{
signals(object)

\S4method{signals}{InTADSig}(object)
}
\arguments{
\item{object}{InTADSig object with signals and genes}
}
\value{
Signals table
}
\description{
This funcion returns the signal values table
}
\examples{
inTadSig <- newSigInTAD(enhSel, enhSelGR, rpkmCountsSel, txsSel)
head(signals(inTadSig))
}
