% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_info.R
\docType{data}
\name{tadGR}
\alias{tadGR}
\title{Genomic coordiantes of topologically associated domains}
\format{a GRanges object}
\usage{
tadGR
}
\value{
NULL, but makes available the dataset
}
\description{
This GRanges object contains the coordinates of TADs
revealed from IMR90 cell line
}
