% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_info.R
\docType{data}
\name{txsSel}
\alias{txsSel}
\title{Genomic coordiantes of genes subset}
\format{a GRanges object}
\usage{
txsSel
}
\value{
NULL, but makes avaialbe the dataset
}
\description{
This GRanges object contains the coordinates of genes
subset from chr15
}
